requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
asyncpg>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6